sidebar: allow turning off tags view by default
With #37065, there is no way to hide the tags from the sidebar
by default when they are not relevant or redundant (e.g. the tab
may already show the file's tags). This can be annyoing especially
when the file has many tags. This patch adds an option to hide
the tags from the sidebar by default (the user can still open
the tags tab manually).

This also reduces one request when opening the sidebar when the
tags are turned off, since all tags don't need to be fetched
anymore.

Signed-off-by: Varun Patil <varunpatil@ucla.edu>